% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpGetAvalMem.R
\name{wpGetAvalMem}
\alias{wpGetAvalMem}
\title{wpGetAvalMem function will return avalible
of the system memory in GB
Tested on Windows 10}
\usage{
wpGetAvalMem()
}
\value{
numeric
}
\description{
wpGetAvalMem function will return avalible
of the system memory in GB
Tested on Windows 10
}
