% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpGetBlocksNeed.R
\name{wpGetBlocksNeed}
\alias{wpGetBlocksNeed}
\title{wpGetBlocksNeed function will return a number of blocks
sugesting for processing raster file. It will take into consideration
number of layers, cells, cores and avalible memory on computer
(not maximum memory but avalible)}
\usage{
wpGetBlocksNeed(x, cores, n = 1, number_type = "numeric")
}
\arguments{
\item{x}{raster}

\item{cores}{number of cores}

\item{n}{parameter to increase requrement of the raster}

\item{number_type}{Will be used to estimate requred memory}
}
\value{
integer
}
\description{
wpGetBlocksNeed function will return a number of blocks
sugesting for processing raster file. It will take into consideration
number of layers, cells, cores and avalible memory on computer
(not maximum memory but avalible)
}
\examples{
wpGetBlocksNeed( x, cores=2, n=1 )

}
