% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpGetIndexesWhichValues.R
\name{wpGetindexesWhichValues}
\alias{wpGetindexesWhichValues}
\title{wpGetindexesWhichValues to get indexes of raster pixels
Script is pirilised and allow to work with big raster}
\usage{
wpGetindexesWhichValues(x, v, cores = NULL, tp = "numeric",
  minblocks = NULL, silent = FALSE)
}
\arguments{
\item{x}{Raster* object}

\item{v}{value of the pixel we would like to get indexes}

\item{cores}{Integer. Number of corest to be used}

\item{tp}{Type of the data to be return. Numberic or}

\item{minblocks}{Integer. Minimum number of blocks. If NULL then it will be calculated automaticly}

\item{silent}{If FALSE then the progress will be shown}
}
\value{
numeric
}
\description{
wpGetindexesWhichValues to get indexes of raster pixels
Script is pirilised and allow to work with big raster
}
\examples{
wpGetindexesWhichValues( x=raster("E:/asm_grid_100m_ccidadminl1.tif"), v=1, cores=4)
}
