% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpGetOS.R
\name{wpGetOS}
\alias{wpGetOS}
\title{wpGetOS function will return a string with OS
of the system
Tested on Windows 10}
\usage{
wpGetOS()
}
\value{
string
}
\description{
wpGetOS function will return a string with OS
of the system
Tested on Windows 10
}
