% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpGetValuesbyInds.R
\name{wpGetValuesbyInds}
\alias{wpGetValuesbyInds}
\title{wpGetValuesbyInds get values of pixels by indexes}
\usage{
wpGetValuesbyInds(x, v, tp = "numeric", cores = NULL, minblocks = NULL,
  nl = 1, silent = TRUE)
}
\arguments{
\item{x}{Raster* object}

\item{v}{indexes of the pixels to be replace with value v}

\item{tp}{Type of the data to be return. Numberic or}

\item{cores}{Integer. Number of cores for parallel calculation}

\item{minblocks}{Integer. Minimum number of blocks. If NULL then it will be calculated automaticly}

\item{nl}{Integer. param to controle min number of blocks during paralisation}

\item{silent}{If FALSE then the progress will be shown}
}
\value{
raster
}
\description{
wpGetValuesbyInds get values of pixels by indexes
}
\examples{
wpGetValuesbyInds( x=raster("E:/asm_grid_100m_ccidadminl1.tif"),v=1, cores=4)
}
