% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpProgressMessage.R
\name{wpProgressMessage}
\alias{wpProgressMessage}
\title{wpProgressMessage function will return progress}
\usage{
wpProgressMessage(x, max = 100, label = NULL)
}
\arguments{
\item{x}{integer}

\item{max}{maximum for progress bvar}

\item{label}{additional text for progress  bar}
}
\value{
character
}
\description{
Tested on Windows 10
}
\examples{
wpProgressMessage( x=10, max = 200, label="Progress message" )
}
