% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpRasterStackCalc.R
\name{wpRasterStackCalc}
\alias{wpRasterStackCalc}
\title{wpRasterStackCalc function to calculate the values 
for a new Raster object from other RasterStack object 
using a formula. Function is paralyzed and suitable for big rasters}
\usage{
wpRasterStackCalc(x, fun = "mean", filename = rasterTmpFile(),
  NAflag = NULL, datatype = NULL, overwrite = TRUE, cores = NULL,
  minblk = NULL, cblk = NULL, silent = TRUE, na.rm)
}
\arguments{
\item{x}{RasterStack object}

\item{fun}{is typically a function that can take a single vector as input}

\item{filename}{File of a new raster file.}

\item{NAflag}{NO data value will be used for a new raster}

\item{datatype}{Type of raster. Avalible are INT1S/INT2S/INT4S/FLT4S/LOG1S/INT1U/INT2U/INT4U/FLT8S}

\item{overwrite}{Overwrite existing file}

\item{cores}{Integer. Number of cores for parallel calculation}

\item{minblk}{Integer. Minimum number of blocks. If NULL then it will be calculated automaticly}

\item{cblk}{Integer. param to controle min number of blocks during paralisation}

\item{silent}{If FALSE then the progress will be shown}

\item{na.rm}{Optional}
}
\value{
Raster* object
}
\description{
wpRasterStackCalc function to calculate the values 
for a new Raster object from other RasterStack object 
using a formula. Function is paralyzed and suitable for big rasters
}
\examples{
wpRasterStackCalc( x, fun='sum')
wpRasterStackCalc( x, fun=function(x){ sum(as.numeric(x))})
wpRasterStackCalc( x, fun=function(x){ (x[1]-x[2])/x[3]})
wpRasterStackCalc( x, fun=function(x){ if (any(!is.na(x))) { return(0) } else { return(NA) }}
			
}
