% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpSetValueWhichindexes.R
\name{wpSetValueWhichindexes}
\alias{wpSetValueWhichindexes}
\title{wpSetValueWhichindexes set pixels with value based on indexes}
\usage{
wpSetValueWhichindexes(x, y, v, filename = rasterTmpFile(), NAflag = NULL,
  datatype = NULL, overwrite = TRUE, cores = NULL, minblk = NULL,
  cblk = NULL, silent = TRUE)
}
\arguments{
\item{x}{Raster* object}

\item{y}{indexes of the pixels to be replace with value v}

\item{v}{value of the pixel we would like to get indexes}

\item{filename}{File of a new raster file.}

\item{NAflag}{NO data value will be used for a new raster}

\item{datatype}{Type of raster. Avalible are INT1S/INT2S/INT4S/FLT4S/LOG1S/INT1U/INT2U/INT4U/FLT8S}

\item{overwrite}{Overwrite existing file}

\item{cores}{Integer. Number of cores for parallel calculation}

\item{minblk}{Integer. Minimum number of blocks. If NULL then it will be calculated automaticly}

\item{cblk}{Integer. param to controle min number of blocks during paralisation}

\item{silent}{If FALSE then the progress will be shown}
}
\value{
raster
}
\description{
wpSetValueWhichindexes set pixels with value based on indexes
}
\examples{
wpSetValueWhichindexes( x=raster("E:/asm_grid_100m_ccidadminl1.tif"), y=1,v=1, cores=4)
}
