% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpTimeDiff.R
\name{wpTimeDiff}
\alias{wpTimeDiff}
\title{wpTimeDiff function will time difference in humon readable format h:m:s}
\usage{
wpTimeDiff(start, end, frm = "hms")
}
\arguments{
\item{start}{time starting}

\item{end}{time endfing}

\item{frm}{format ooutput}
}
\value{
character
}
\description{
wpTimeDiff function will time difference in humon readable format h:m:s
}
\examples{
wpTimeDiff( start=Sys.time(), end=Sys.time(), frm="hms" )

}
