% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpZonalStatistics.R
\name{wpZonalStatistics}
\alias{wpZonalStatistics}
\title{wpZonalStatistics function compute zonal statistics. That is,
cross-tabulate the values of a Raster* object
based on a "zones" RasterLayer. NA values are removed.
Function uses DoParallel library to work with a big raster data}
\usage{
wpZonalStatistics(x, y, fun = "mean", cores = NULL, minblk = NULL,
  silent = TRUE)
}
\arguments{
\item{x}{Raster* object}

\item{y}{RasterLayer object with codes representing zones}

\item{fun}{The function to be applied. Either as character: 'mean', 'min', 'max' and 'sum'}

\item{cores}{Integer. Number of cores for parallel calculation}

\item{minblk}{Integer. Minimum number of blocks}

\item{silent}{If FALSE then the progress will be shown}
}
\value{
A data.frame with a value for each zone (unique value in zones)
}
\description{
wpZonalStatistics function compute zonal statistics. That is,
cross-tabulate the values of a Raster* object
based on a "zones" RasterLayer. NA values are removed.
Function uses DoParallel library to work with a big raster data
}
\examples{
wpZonalStatistics( x=rasterObj1, y=rasterObj2, cores=2, minblk=4  )

}
